move actual testing into helper file to make it easier to add a benchmark file
